feat(images): extract utils functions, remove loading='lazy'
