+++ packages/@ourworldindata/utils/src/image.ts
@@ -1,0 +1,60 @@
+/*
+Common utlities for deriving properties from image metadata.
+*/
+
+// This is the JSON we get from Google's API before remapping the keys to be consistent with the rest of our interfaces
+export interface GDriveImageMetadata {
+    name: string // -> filename
+    modifiedTime: string // -> updatedAt e.g. "2023-01-11T19:45:27.000Z"
+    id: string // -> googleId e.g. "1dfArzg3JrAJupVl4YyJpb2FOnBn4irPX"
+    description?: string // -> defaultAlt
+    imageMediaMetadata?: {
+        width?: number // -> originalWidth
+    }
+}
+
+export interface ImageMetadata {
+    googleId: string
+    filename: string
+    defaultAlt: string
+    // MySQL Date objects round to the nearest second, whereas Google includes milliseconds
+    // so we store as an epoch to avoid any conversion issues
+    updatedAt: number
+    originalWidth: number
+}
+
+export function getSizes(
+    originalWidth: ImageMetadata["originalWidth"]
+): number[] | undefined {
+    // ensure a thumbnail is generated
+    const widths = [100]
+    // start at 350 and go up by 500 to a max of 1350 before we just show the original image
+    let width = 350
+    while (width < originalWidth && width <= 1350) {
+        widths.push(width)
+        width += 500
+    }
+    widths.push(originalWidth)
+    return widths
+}
+
+export function generateSrcSet(
+    sizes: number[],
+    filename: ImageMetadata["filename"]
+): string {
+    return sizes
+        .map(
+            (size) =>
+                `images/${filename.slice(
+                    0,
+                    filename.indexOf(".")
+                )}_${size}.webp ${size}w`
+        )
+        .join(", ")
+}
+
+export function getFilenameWithoutExtension(
+    filename: ImageMetadata["filename"]
+): string {
+    return filename.slice(0, filename.indexOf("."))
+}
